requirements.txt
Django~=3.2.6
pip~=20.1.1
pytz~=2019.3
wheel~=0.34.2
sqlparse~=0.4.1
PyYAML~=5.3.1
pylint~=2.10.2
environ~=1.0
dj-config-url~=0.1.1
psycopg2-binary~=2.9.1
envparse~=0.1.6
django-crispy-forms~=1.12.0
django-staticfiles~=1.2.1
djangorestframework~=3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
factory-boy==3.2.0
halo==0.0.31
django-filter==2.4.0
